+++ python/tvm/contrib/tar.py
@@ -99,7 +99,7 @@ def normalize_file_list_by_unpacking_tars(temp, file_list):
 
     Parameters
     ----------
-    temp: tvm.contrib.utils.TempDirectory
+    temp: tvm.support.utils.TempDirectory
         A temp dir to hold the untared files.
 
     file_list: List[str]
