+++ tests/python/runtime/test_runtime_measure.py
@@ -20,7 +20,7 @@ import time
 
 import tvm
 from tvm import te
-from tvm.contrib.utils import tempdir
+from tvm.support.utils import tempdir
 from tvm.runtime.module import BenchmarkResult
