+++ web/tests/python/webgpu_rpc_test.py
@@ -24,7 +24,8 @@ import numpy as np
 
 import tvm
 from tvm import rpc, te
-from tvm.contrib import tvmjs, utils
+from tvm.contrib import tvmjs
+from tvm.support import utils
 
 proxy_host = "127.0.0.1"
 proxy_port = 9090
